+++ modules/apps/29-fee/client/cli/query.go
@@ -120,7 +120,7 @@ func GetCmdTotalRecvFees() *cobra.Command {
 				return err
 			}
 
-			packetID := channeltypes.NewPacketId(channelID, portID, seq)
+			packetID := channeltypes.NewPacketId(portID, channelID, seq)
 
 			if err := packetID.Validate(); err != nil {
 				return err
@@ -166,7 +166,7 @@ func GetCmdTotalAckFees() *cobra.Command {
 				return err
 			}
 
-			packetID := channeltypes.NewPacketId(channelID, portID, seq)
+			packetID := channeltypes.NewPacketId(portID, channelID, seq)
 
 			if err := packetID.Validate(); err != nil {
 				return err
@@ -212,7 +212,7 @@ func GetCmdTotalTimeoutFees() *cobra.Command {
 				return err
 			}
 
-			packetID := channeltypes.NewPacketId(channelID, portID, seq)
+			packetID := channeltypes.NewPacketId(portID, channelID, seq)
 
 			if err := packetID.Validate(); err != nil {
 				return err