+++ modules/apps/29-fee/keeper/msg_server.go
@@ -38,8 +38,8 @@ func (k Keeper) PayPacketFee(goCtx context.Context, msg *types.MsgPayPacketFee) 
 	}
 
 	packetID := channeltypes.NewPacketId(
-		msg.SourceChannelId,
 		msg.SourcePortId,
+		msg.SourceChannelId,
 		sequence,
 	)